flytt uthenting av utbetalingsperioder inn i oppsummeringsboks
